requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.10
pydantic==2.10.3
openai==1.12.0